requirements.txt
dify_plugin>=0.2.0,<0.3.0
mysql-connector-python==9.3.0
